minor memory optims: use subgraph id instead of pointer value
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@2408 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/include/tulip/DoubleProperty.h
@@ -50,9 +50,9 @@ public :
 			      PredefinedMetaValueCalculator edgeCalc = AVG_CALC);
 
 private:
-  TLP_HASH_MAP<unsigned long, double> maxN,minN,maxE,minE;
-  TLP_HASH_MAP<unsigned long, bool> minMaxOkNode;
-  TLP_HASH_MAP<unsigned long, bool> minMaxOkEdge;
+  TLP_HASH_MAP<unsigned int, double> maxN,minN,maxE,minE;
+  TLP_HASH_MAP<unsigned int, bool> minMaxOkNode;
+  TLP_HASH_MAP<unsigned int, bool> minMaxOkEdge;
   void computeMinMaxNode(Graph *sg=0);
   void computeMinMaxEdge(Graph *sg=0);
 };
